+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 ### Added
 - Method in sfrRecord manager to check for reasonable publication dates for editions
+- Tests for publicationDateCheck method in sfrRecord manager
 
 ## 2021-11-22 -- v0.9.5
 ### Added